Faster getPartitionAssignment
