--- hazelcast/src/main/java/com/hazelcast/jet/impl/execution/init/IntArrayWithPosition.java
@@ -1,23 +1,0 @@
-package com.hazelcast.jet.impl.execution.init;
-
-import java.util.Arrays;
-
-class IntArrayWithPosition {
-    private final int[] array;
-    private int pos;
-
-    IntArrayWithPosition(int size) {
-        array = new int[size];
-    }
-
-    void add(int i) {
-        array[pos++] = i;
-    }
-
-    int[] getFilledElements() {
-        if (pos == array.length) {
-            return array;
-        }
-        return Arrays.copyOfRange(array, 0, pos);
-    }
-}
